fixed sql_commands bug
I fixed a bug with both sql commands in the sql_commands.py file where I needed to put the full path to the db file in order to be able to connect to the db. I made it so I just put the relative path which means it can be used on any computer the files are downloaded in.